backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9

# ── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ── Redis / Caching ─────────────────────────────────────────
redis==5.1.1

# ── Task Queue ───────────────────────────────────────────────
celery==5.4.0

# ── Auth ─────────────────────────────────────────────────────
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0

# ── AI / NLP ─────────────────────────────────────────────────
openai==1.47.0
anthropic==0.34.2
tiktoken==0.7.0

# ── Resume Parsing ───────────────────────────────────────────
PyMuPDF==1.24.10
python-docx==1.1.2

# ── Data Validation ──────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ── HTTP Client ──────────────────────────────────────────────
httpx==0.27.2

# ── Logging ──────────────────────────────────────────────────
structlog==24.4.0

# ── Testing ──────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# ── PDF Export ───────────────────────────────────────────────
reportlab==4.2.2

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
